Put parseArgs into the one location where it is used.
